math_instruction_tuning/requirements.txt
accelerate==0.27.2
beautifulsoup4==4.11.1
bitsandbytes
fraction
bleach==5.0.1
blessed==1.20.0
cached_path==1.6.3
cachetools==4.2.4
Cython==0.29.32
datasets==2.16.1
evaluate==0.4.0
fastBPE==0.1.0
fasteners==0.18
fastjsonschema==2.16.2
future==0.18.2
httptools==0.6.1
huggingface-hub==0.23.5
jsonable==0.3.1
jsonlines==4.0.0
jsonschema==3.2.0
llvmlite==0.41.1
lm-format-enforcer==0.9.8
lxml==4.9.1
matplotlib
multidict==6.0.2
numpy==1.22.0
oauthlib==3.1.0
omegaconf==2.3.0
pandas==2.0.3
peft==0.13.0
pure-eval==0.2.2
pyarrow==9.0.0
PyYAML==6.0
QtPy==2.2.1
regex==2022.10.31
requests
responses==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==0.24.0
scipy==1.10.0
Send2Trash==1.8.0
sentencepiece==0.1.91
sentry-sdk==1.10.1
seqeval==1.2.2
simplejson==3.17.2
smart-open==6.3.0
sympy==1.12
tensorboard==2.13.0
tokenizers==0.19.1
torch==2.2.1
torchaudio==2.2.1
torchmetrics==0.11.3
torchvision==0.17.1
tqdm
transformers==4.40.1
triton==2.2.0
vllm==0.4.1
wandb==0.13.5
xformers==0.0.25
yarl==1.8.1  
